Fix cpu ci check failures

+++ lib/task-spec/include/task-spec/forward_tensor_source.h
@@ -14,7 +14,7 @@ public:
   void reset();
 
 private:
-  static size_t next_available_forward_tensor_id;
+  static int next_available_forward_tensor_id;
 };
 
 } // namespace FlexFlow
